fix(schema): resolve project scaling issue by dropping embedded `profile.project` id array
